+++ src/podofo/main/PdfMemDocument.h
@@ -141,7 +141,7 @@ public:
      */
     void SetEncrypted(const std::string_view& userPassword, const std::string_view& ownerPassword,
         PdfPermissions protection = PdfPermissions::Default,
-        PdfEncryptionAlgorithm algorithm = PdfEncryptionAlgorithm::AESV2,
+        PdfEncryptionAlgorithm algorithm = PdfEncryptionAlgorithm::AESV3R6,
         PdfKeyLength keyLength = PdfKeyLength::Unknown);
 
     /** Encrypt the document during writing using a PdfEncrypt object